+++ src/main/java/cz/cvut/fel/ear/hamrazec/dormitory/dao/GenericDao.java
@@ -16,7 +16,7 @@ public interface GenericDao<T> {
      * @param id Identifier
      * @return Entity instance or {@code null} if no such instance exists
      */
-    T find(Integer id);
+    T find(Long id);
 
     /**
      * Finds all instances of the specified class.
